pyproject.toml
[tool.poetry]
name = "python-fastapi-poc"
version = "0.1.0"
description = "FastAPI Sequence management proof of concept"
authors = []
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.0"
uvicorn = { extras = ["standard"], version = "^0.32.0" }
sqlalchemy = "^2.0.0"
alembic = "^1.14.0"
psycopg2-binary = "^2.9.0"
pydantic = "^2.10.0"
pydantic-settings = "^2.6.0"
prometheus-fastapi-instrumentator = "^7.1"
python-logging-loki = "^0.3.1"
python-json-logger = "^3.2"
python-jose = {extras = ["cryptography"], version = "^3.5.0"}
bcrypt = "^5.0.0"
python-multipart = "^0.0.22"

[tool.poetry.group.dev.dependencies]
ruff = "^0.8.0"
pytest = "^8.3.0"
pytest-cov = "^6.0.0"
httpx = "^0.28.0"
allure-pytest = "^2.13.0"
testcontainers = {extras = ["postgres"], version = "^4.14.1"}
basedpyright = "^1.38.2"
playwright = "^1.58.0"

[tool.ruff]
line-length = 88
exclude = ["alembic"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.basedpyright]
typeCheckingMode = "strict"
pythonVersion = "3.12"
reportMissingTypeStubs = "none"
reportUntypedFunctionDecorator = "none"
reportUntypedClassDecorator = "none"
exclude = ["alembic", ".venv"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --ignore=tests/perf --ignore=tests/e2e"
markers = [
    "perf: performance/timing tests (run with: just perf)",
    "e2e: end-to-end tests against a live running stack (run with: just e2e)",
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
